requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
fastapi==0.106.0
h11==0.14.0
httptools==0.6.1
idna==3.6
joblib==1.3.2
numpy==1.26.2
passlib==1.7.4
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
Sastrawi==1.0.1
scikit-learn==1.3.2
scipy==1.11.4
sniffio==1.3.0
snowflake-id==0.0.5
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
